Apresentação + Migration + Ajustes

+++ BNP.Movimentacoes.Aplicacao/DTOs/MovimentoManualDto.cs
@@ -2,11 +2,11 @@
 {
     public class MovimentoManualDto
     {
-        public decimal Mes { get; set; }
-        public decimal Ano { get; set; }
+        public int Mes { get; set; }
+        public int Ano { get; set; }
         public string CodProduto { get; set; }
         public string? NomeProduto { get; set; }
-        public decimal NumLancamento { get; set; }
+        public long NumLancamento { get; set; }
         public string Descricao { get; set; }
         public decimal Valor { get; set; }
     }
